# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Paskihub API Documentation",
        "title": "Paskihub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3010",
    "basePath": "/",
    "paths": {
...
        "/api/v1/eo/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve current Event Organizer's profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Profile"
                ],
                "summary": "Get EO profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.EOProfileRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/eo/profile/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Update password for the current Event Organizer account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Profile"
                ],
                "summary": "Update EO password",
                "parameters": [
                    {
                        "description": "Update Password Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EOUpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/eo/staff": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of staff accounts associated with the current Event Organizer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Staff"
                ],
                "summary": "Get list of staff",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.EOStaffRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new staff account under the current Event Organizer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Staff"
                ],
                "summary": "Create a new staff account",
                "parameters": [
                    {
                        "description": "Staff Create Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EOStaffCreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/eo/staff/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently removes a staff account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Staff"
                ],
                "summary": "Delete a staff account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Staff ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/eo/staff/{id}/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the password for a specific staff account (Reset by EO)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Staff"
                ],
                "summary": "Reset staff password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Staff ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reset Password Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EOStaffResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EOProfileRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "dto.EOStaffCreateReq": {
            "type": "object",
            "required": [
                "confirm_password",
                "email",
                "password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.EOStaffRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.EOStaffResetPasswordReq": {
            "type": "object",
            "required": [
                "confirm_password",
                "password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "dto.EOUpdatePasswordReq": {
            "type": "object",
            "required": [
                "confirm_password",
                "new_password",
                "old_password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/response.Status"
                }
            }
        },
        "response.Status": {
            "type": "string",
            "enum": [
                "success",
                "fail",
                "error"
            ],
            "x-enum-varnames": [
                "Success",
                "Fail",
                "Error"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for all endpoints. Format: Key {api_key}. Example: Key abc123",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}